init covers grid
