+++ frontend/src/components/Handling.js
@@ -1,6 +1,4 @@
 import React, { Component } from 'react';
-import { Link } from "react-router-dom"
-
 
 class LoadingPage extends Component {
     render(){
