+++ server/server.py
@@ -55,6 +55,7 @@ def goodreads_process_reviews(xml_content):
         }
         reviews.append(review)
 
+    reviews.reverse()
     return reviews
 
 
@@ -73,8 +74,8 @@ def get_year_data(year, xml_content):
             'total_pages' : sum(review['num_pages'] for review in reviews),
             'avg_pages' : round(sum(review['num_pages'] for review in reviews) / len(reviews)),
             'avg_rating' : round((sum(review['your_rating'] for review in reviews) / len(reviews)), 2),
-            'first_book' : reviews[-1],
-            'last_book': reviews[0],
+            'first_book' : reviews[0],
+            'last_book': reviews[-1],
             'shortest_book' : next(review for review in sorted_by_pages if review['num_pages'] > 0),  # books with 0 pages (ex. webcomics)
             'longest_book' : sorted_by_pages[-1],
             'highest_rated_book' : sorted_by_rating[0],
